Split Image components into files
