+++ app/containers/message/Components/Attachments/Image/Button.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+import { useTheme } from '../../../../../theme';
+import Touchable from '../../../Touchable';
+import styles from '../../../styles';
+
+interface IMessageButton {
+	children: React.ReactElement;
+	disabled?: boolean;
+	onPress: () => void;
+}
+
+export const Button = React.memo(({ children, onPress, disabled }: IMessageButton) => {
+	const { colors } = useTheme();
+	return (
+		<Touchable
+			disabled={disabled}
+			onPress={onPress}
+			style={styles.imageContainer}
+			background={Touchable.Ripple(colors.surfaceNeutral)}>
+			{children}
+		</Touchable>
+	);
+});
