+++ app/containers/message/Components/Attachments/Image/index.tsx
@@ -1,0 +1,4 @@
+import Container from './Container';
+
+export * from './Image';
+export default Container;
